requirements.txt
pydantic~=2.7.3
azure-functions==1.19.0
pydantic-settings==2.3.1
